refactor: removed non-essentials from utils-maps. deleted math.ts, strongly reduced engine.shapes.ts, deleted colors.ts. (Note: recomending chroma.js and glMatrix as replacements, respectively).
