+++ projects/demo-maps/src/app/route-components/route-example-custom-layers/route-map4.component.html
@@ -5,14 +5,6 @@
 <clr-vertical-nav [clrVerticalNavCollapsible]="true" class="right" [clr-nav-level]="2">
 
   <clr-vertical-nav-group [clrVerticalNavGroupExpanded]="true" class="layers">
-    <clr-icon shape="clock" title="Performance" clrVerticalNavIcon></clr-icon>
-    Performance
-    <clr-vertical-nav-group-children class="padding title-ellipsis">
-        <ukis-performance></ukis-performance>
-    </clr-vertical-nav-group-children>
-  </clr-vertical-nav-group>
-
-  <clr-vertical-nav-group [clrVerticalNavGroupExpanded]="true" class="layers">
     <clr-icon shape="layers" clrVerticalNavIcon title="layers"></clr-icon>
     Layers
     <clr-vertical-nav-group-children class="padding title-ellipsis">
