+++ projects/utils-browser/src/public-api.ts
@@ -4,4 +4,3 @@
 
 export * from './lib/download/download';
 export * from './lib/layout/layout';
-export * from './lib/layout/colors';
