+++ projects/utils-maps/src/public-api.ts
@@ -10,4 +10,3 @@ export * from './lib/webgl/webgl';
 export * from './lib/webgl/engine.core';
 export * from './lib/webgl/engine.helpers';
 export * from './lib/webgl/engine.shapes';
-export * from './lib/webgl/math';
